#14 - using PageView to swipe through song lyrics
